rag-backend/requirements.txt
fastapi
uvicorn
langchain
langchain_openai
uuid
langchain_text_splitters
langchain_community
python-dotenv
pypdf
psycopg2-binary
pgvector
SQLAlchemy